pyproject.toml
[project]
name = "final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.1"
dependencies = [
    "datasets>=3.5.0",
    "gradio>=5.27.0",
    "librosa>=0.11.0",
    "matplotlib>=3.10.1",
    "midi2audio>=0.1.1",
    "midiutil>=1.2.1",
    "mido>=1.3.3",
    "music21>=9.3.0",
    "numpy>=2.2.5",
    "pretty-midi>=0.2.10",
    "pyaudio>=0.2.14",
    "pygame>=2.6.1",
    "python-dotenv>=1.1.0",
    "setuptools>=79.0.1",
    "soundfile>=0.13.1",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.3",
    "wandb>=0.19.10",
]
